backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (MongoDB)
pymongo==4.16.0
motor==3.7.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Blockchain Integration
web3==6.15.0
python-dotenv==1.0.0

# IPFS Integration
requests==2.31.0
